Change seed value type from Int to Long
Remove redundant .toLong casts in McRunner.runSingle and
WorldInit.initialize — seed flows as Long end-to-end.

+++ src/main/scala/sfc/init/WorldInit.scala
@@ -15,8 +15,8 @@ import scala.util.Random
 object WorldInit:
 
   /** Initialize a complete simulation world from a seed. */
-  def initialize(seed: Int)(using p: SimParams): InitResult =
-    val rng = new Random(seed.toLong)
+  def initialize(seed: Long)(using p: SimParams): InitResult =
+    val rng = new Random(seed)
 
     // --- Firms ---
     val firms = FirmInit.create(rng)
